refactor(ckpt): refactor bootstrap as trait
