+++ src/ws-ckpt/src/crates/daemon/src/backends/btrfs_loop.rs
@@ -8,9 +8,10 @@ use tokio::process::Command;
 use tracing::{error, info, warn};
 
 use ws_ckpt_common::backend::*;
-use ws_ckpt_common::{DiffEntry, WorkspaceInfo, SNAPSHOTS_DIR};
+use ws_ckpt_common::{DaemonConfig, DiffEntry, WorkspaceInfo, SNAPSHOTS_DIR};
 
 use super::btrfs_common;
+use crate::util::{is_mounted, run_command, run_command_checked};
 use btrfs_common::resolve_symlink_path;
 
 pub struct BtrfsLoopBackend {
@@ -401,5 +402,396 @@ impl StorageBackend for BtrfsLoopBackend {
                     self.mount_path.display()
                 )
             })
+    }
+
+    async fn bootstrap(&self, config: &DaemonConfig) -> anyhow::Result<()> {
+        btrfs_common::ensure_btrfs_support()
+            .await
+            .context("btrfs kernel support is required")?;
+
+        let img_path = &config.img_path;
+        let img_dir = derive_img_dir(img_path)
+            .context("Failed to derive image directory from config.img_path")?;
+
+        tokio::fs::create_dir_all(&img_dir)
+            .await
+            .context("Failed to create ws-ckpt data directory")?;
+
+        // A newly created image already matches target; only reconcile pre-existing ones.
+        let img_existed_before = tokio::fs::metadata(img_path).await.is_ok();
+        if !img_existed_before {
+            create_sparse_image(config, &img_dir).await?;
+        }
+
+        tokio::fs::create_dir_all(&config.mount_path)
+            .await
+            .context("Failed to create mount point directory")?;
+
+        let mount_path_str = config.mount_path.to_string_lossy().to_string();
+        if !is_mounted(&mount_path_str).await? {
+            let loop_device = run_command("losetup", &["--find", "--show", &config.img_path])
+                .await
+                .context("Failed to setup loop device")?;
+            let loop_device = loop_device.trim().to_string();
+            run_command_checked("mount", &[&loop_device, &mount_path_str])
+                .await
+                .context("Failed to mount btrfs image")?;
+            info!("Mounted {} at {}", loop_device, mount_path_str);
+        } else {
+            info!("Already mounted at {:?}", config.mount_path);
+        }
+
+        if img_existed_before {
+            if let Err(e) = reconcile_img_size(config).await {
+                warn!(
+                    "Failed to reconcile btrfs image size: {:#}. Continuing with current size.",
+                    e
+                );
+            }
+        }
+
+        let snapshots_dir = config.mount_path.join(SNAPSHOTS_DIR);
+        tokio::fs::create_dir_all(&snapshots_dir)
+            .await
+            .context("Failed to create snapshots directory")?;
+
+        info!("BtrfsLoop bootstrap complete");
+        Ok(())
+    }
+}
+
+// ────────────────────────────────────────────────────────────────────────────
+// BtrfsLoop-specific bootstrap helpers
+// ────────────────────────────────────────────────────────────────────────────
+
+/// Create a sparse image file sized by `min(img_size GB, total * img_max_percent%)`,
+/// degrading to `avail * img_max_percent%` if target exceeds host avail. Then mkfs.btrfs.
+async fn create_sparse_image(config: &DaemonConfig, img_dir: &str) -> anyhow::Result<()> {
+    let img_path = &config.img_path;
+    // `-P` forces POSIX single-line output; long device names must not wrap.
+    let df_output = run_command("df", &["-P", "-B1", img_dir])
+        .await
+        .context("Failed to get partition info")?;
+    let total = parse_df_total(&df_output).context("Failed to parse df total")?;
+    let avail = parse_df_available(&df_output).context("Failed to parse df output")?;
+
+    let target = compute_target_size(config.img_size, config.img_max_percent, total);
+
+    const GB: f64 = 1024.0 * 1024.0 * 1024.0;
+    let img_size = if target > avail {
+        let degraded = (avail as f64 * config.img_max_percent / 100.0) as u64;
+        warn!(
+            "Target image size {:.1} GB exceeds available {:.1} GB. Degrading to {:.1} GB ({}% of available).",
+            target as f64 / GB,
+            avail as f64 / GB,
+            degraded as f64 / GB,
+            config.img_max_percent,
+        );
+        degraded
+    } else {
+        target
+    };
+    info!(
+        "Creating sparse image {} bytes ({:.1} GB), total {:.1} GB, avail {:.1} GB",
+        img_size,
+        img_size as f64 / GB,
+        total as f64 / GB,
+        avail as f64 / GB,
+    );
+
+    run_command_checked("truncate", &["-s", &img_size.to_string(), img_path])
+        .await
+        .context("Failed to create sparse image file")?;
+    run_command_checked("mkfs.btrfs", &["-f", img_path])
+        .await
+        .context("Failed to format btrfs image")?;
+    Ok(())
+}
+
+/// img_path must be absolute with a non-empty parent dir.
+fn derive_img_dir(img_path: &str) -> anyhow::Result<String> {
+    let parent = Path::new(img_path)
+        .parent()
+        .filter(|p| !p.as_os_str().is_empty())
+        .with_context(|| {
+            format!(
+                "Invalid img_path {:?}: must be an absolute path containing a parent directory",
+                img_path
+            )
+        })?;
+    Ok(parent.to_string_lossy().to_string())
+}
+
+/// Smaller of: absolute cap (`img_size_gb * GiB`) and percentage cap (`total * img_max_percent%`).
+fn compute_target_size(img_size_gb: u64, img_max_percent: f64, total_bytes: u64) -> u64 {
+    let by_size = img_size_gb.saturating_mul(1024 * 1024 * 1024);
+    let by_percent = (total_bytes as f64 * img_max_percent / 100.0) as u64;
+    std::cmp::min(by_size, by_percent)
+}
+
+/// Reconcile on-disk img size against computed target.
+///
+/// - Equal: no-op (still runs a final `btrfs fs resize max` for self-healing).
+/// - Less : grow in place via truncate + `losetup -c`, guarded by host avail.
+/// - Greater: shrink via btrfs resize + unmount + truncate + remount, guarded by `fuser -m`.
+async fn reconcile_img_size(config: &DaemonConfig) -> anyhow::Result<()> {
+    let img_path = &config.img_path;
+    let mount_path_str = config.mount_path.to_string_lossy().to_string();
+    let current = tokio::fs::metadata(img_path)
+        .await
+        .with_context(|| format!("Failed to stat image {}", img_path))?
+        .len();
+
+    let img_dir = derive_img_dir(img_path)
+        .context("Failed to derive image directory from config.img_path")?;
+    let df_output = run_command("df", &["-P", "-B1", &img_dir])
+        .await
+        .context("Failed to get host partition info for reconcile")?;
+    let total = parse_df_total(&df_output).context("Failed to parse df total")?;
+    let avail = parse_df_available(&df_output).context("Failed to parse df available")?;
+
+    let target = compute_target_size(config.img_size, config.img_max_percent, total);
+
+    // Track whether we actually grew; only log "grown to" after the final fs resize succeeds.
+    let mut grew_to: Option<u64> = None;
+
+    match current.cmp(&target) {
+        std::cmp::Ordering::Equal => {
+            info!("Btrfs image size already matches target: {} bytes", current);
+        }
+        std::cmp::Ordering::Less => {
+            let needed = target - current;
+            if needed > avail {
+                warn!(
+                    "Cannot grow btrfs image: need {} more bytes (target {}) but avail {}. Keeping current ({} bytes).",
+                    needed, target, avail, current,
+                );
+                return Ok(());
+            }
+            info!("Growing btrfs image: {} -> {} bytes", current, target);
+            run_command_checked("truncate", &["-s", &target.to_string(), img_path])
+                .await
+                .context("Failed to grow sparse image file")?;
+            let loop_device = find_loop_device_for(img_path)
+                .await
+                .context("Failed to locate loop device for image")?;
+            run_command_checked("losetup", &["-c", &loop_device])
+                .await
+                .context("Failed to refresh loop device capacity")?;
+            // Final `btrfs fs resize max` runs below; on failure state is recoverable.
+            grew_to = Some(target);
+        }
+        std::cmp::Ordering::Greater => {
+            if let Some(pids) = check_mount_busy(&mount_path_str).await {
+                warn!(
+                    "Cannot shrink btrfs image: mount {} in use by PIDs [{}]. Keeping current ({} bytes).",
+                    mount_path_str, pids, current,
+                );
+                return Ok(());
+            }
+            warn!("Shrinking btrfs image: {} -> {} bytes", current, target);
+            // Shrink fs first — fails cleanly if used bytes exceed target, no data loss.
+            run_command_checked(
+                "btrfs",
+                &["filesystem", "resize", &target.to_string(), &mount_path_str],
+            )
+            .await
+            .context("Failed to shrink btrfs filesystem (data may exceed new size)")?;
+            let loop_device = find_loop_device_for(img_path)
+                .await
+                .context("Failed to locate loop device for image")?;
+            run_command_checked("umount", &[&mount_path_str])
+                .await
+                .context("Failed to unmount for image shrink")?;
+            run_command_checked("losetup", &["-d", &loop_device])
+                .await
+                .context("Failed to detach loop device")?;
+            run_command_checked("truncate", &["-s", &target.to_string(), img_path])
+                .await
+                .context("Failed to truncate image file")?;
+            let new_loop = run_command("losetup", &["--find", "--show", img_path])
+                .await
+                .context("Failed to reattach loop device")?;
+            let new_loop = new_loop.trim().to_string();
+            run_command_checked("mount", &[&new_loop, &mount_path_str])
+                .await
+                .context("Failed to remount after image shrink")?;
+            info!("Btrfs image shrunk to {} bytes", target);
+        }
+    }
+
+    // Single exit point: syncs fs size to loop capacity. Also self-heals a previous
+    // half-done grow (truncate+losetup -c succeeded but btrfs resize died).
+    run_command_checked("btrfs", &["filesystem", "resize", "max", &mount_path_str])
+        .await
+        .context("Failed to sync btrfs filesystem size to loop capacity")?;
+    if let Some(t) = grew_to {
+        info!("Btrfs image grown to {} bytes", t);
+    }
+    Ok(())
+}
+
+/// Advisory check via `fuser -m`; returns PIDs occupying the mount or None.
+/// Not routed through `run_command` because fuser exits 1 on the common
+/// "nothing using it" path, which `run_command` would treat as hard failure.
+async fn check_mount_busy(mount_path: &str) -> Option<String> {
+    let output = Command::new("fuser")
+        .env("LC_ALL", "C")
+        .env("LANG", "C")
+        .args(["-m", mount_path])
+        .output()
+        .await
+        .ok()?;
+    let pids = String::from_utf8_lossy(&output.stdout).trim().to_string();
+    if pids.is_empty() {
+        None
+    } else {
+        Some(pids)
+    }
+}
+
+/// Parse `losetup -j <img>` and return the backing loop device.
+async fn find_loop_device_for(img_path: &str) -> anyhow::Result<String> {
+    let out = run_command("losetup", &["-j", img_path])
+        .await
+        .context("Failed to run `losetup -j`")?;
+    parse_losetup_j(&out, img_path)
+}
+
+fn parse_losetup_j(out: &str, img_path: &str) -> anyhow::Result<String> {
+    // Expected: "/dev/loop0: [2049]:12345 (/var/lib/ws-ckpt/btrfs-data.img)"
+    let line = out
+        .lines()
+        .next()
+        .filter(|s| !s.trim().is_empty())
+        .with_context(|| format!("No loop device currently backs image {}", img_path))?;
+    let device = line
+        .split(':')
+        .next()
+        .with_context(|| format!("Cannot parse losetup output: {}", line))?;
+    Ok(device.trim().to_string())
+}
+
+/// Parse `Available` column (index 3) from `df -B1` output.
+fn parse_df_available(output: &str) -> anyhow::Result<u64> {
+    let line = output
+        .lines()
+        .nth(1)
+        .context("df output has no data line")?;
+    let avail_str = line
+        .split_whitespace()
+        .nth(3)
+        .context("df output missing available column")?;
+    avail_str
+        .parse::<u64>()
+        .context("Failed to parse available size from df output")
+}
+
+/// Parse `1B-blocks` (total) column (index 1) from `df -B1` output.
+fn parse_df_total(output: &str) -> anyhow::Result<u64> {
+    let line = output
+        .lines()
+        .nth(1)
+        .context("df output has no data line")?;
+    let total_str = line
+        .split_whitespace()
+        .nth(1)
+        .context("df output missing total column")?;
+    total_str
+        .parse::<u64>()
+        .context("Failed to parse total size from df output")
+}
+
+#[cfg(test)]
+mod tests {
+    use super::{
+        compute_target_size, derive_img_dir, parse_df_available, parse_df_total, parse_losetup_j,
+    };
+
+    const GB: u64 = 1024 * 1024 * 1024;
+
+    #[test]
+    fn parses_available_column_from_df_b1() {
+        let out = "Filesystem     1B-blocks        Used   Available Use% Mounted on\n\
+                   /dev/sda1    107374182400 32212254720 75161927680  30% /data\n";
+        assert_eq!(parse_df_available(out).unwrap(), 75_161_927_680u64);
+    }
+
+    #[test]
+    fn returns_err_on_missing_data_line() {
+        let out = "Filesystem 1B-blocks Used Available Use% Mounted on\n";
+        assert!(parse_df_available(out).is_err());
+    }
+
+    #[test]
+    fn returns_err_on_non_numeric_available() {
+        let out = "Filesystem 1B-blocks Used Available Use% Mounted on\n\
+                   /dev/sda1 100 10 NaN 10% /data\n";
+        assert!(parse_df_available(out).is_err());
+    }
+
+    #[test]
+    fn parses_loop_device_from_losetup_j() {
+        let out = "/dev/loop0: [2049]:12345 (/var/lib/ws-ckpt/btrfs-data.img)\n";
+        assert_eq!(
+            parse_losetup_j(out, "/var/lib/ws-ckpt/btrfs-data.img").unwrap(),
+            "/dev/loop0"
+        );
+    }
+
+    #[test]
+    fn parse_losetup_j_returns_err_on_empty() {
+        assert!(parse_losetup_j("", "/var/lib/ws-ckpt/btrfs-data.img").is_err());
+        assert!(parse_losetup_j("\n", "/var/lib/ws-ckpt/btrfs-data.img").is_err());
+    }
+
+    #[test]
+    fn parses_total_column_from_df_b1() {
+        let out = "Filesystem     1B-blocks        Used   Available Use% Mounted on\n\
+                   /dev/sda1    107374182400 32212254720 75161927680  30% /data\n";
+        assert_eq!(parse_df_total(out).unwrap(), 107_374_182_400u64);
+    }
+
+    #[test]
+    fn compute_target_picks_size_cap_when_smaller() {
+        let got = compute_target_size(30, 50.0, 100 * GB);
+        assert_eq!(got, 30 * GB);
+    }
+
+    #[test]
+    fn compute_target_picks_percent_cap_when_smaller() {
+        let got = compute_target_size(30, 50.0, 40 * GB);
+        assert_eq!(got, 20 * GB);
+    }
+
+    #[test]
+    fn compute_target_handles_equal_caps() {
+        let got = compute_target_size(30, 50.0, 60 * GB);
+        assert_eq!(got, 30 * GB);
+    }
+
+    #[test]
+    fn compute_target_saturates_on_huge_img_size() {
+        let got = compute_target_size(u64::MAX / GB, 50.0, 100 * GB);
+        assert_eq!(got, 50 * GB);
+    }
+
+    #[test]
+    fn derive_img_dir_returns_parent_for_absolute_path() {
+        assert_eq!(
+            derive_img_dir("/var/lib/ws-ckpt/btrfs-data.img").unwrap(),
+            "/var/lib/ws-ckpt"
+        );
+    }
+
+    #[test]
+    fn derive_img_dir_rejects_bare_filename() {
+        assert!(derive_img_dir("data.img").is_err());
+    }
+
+    #[test]
+    fn derive_img_dir_rejects_root_and_empty() {
+        assert!(derive_img_dir("/").is_err());
+        assert!(derive_img_dir("").is_err());
     }
 }
